CCEffect - Fix signed / unsigned comparison warnings when dealing with pass counts and indices
Since passes are held internally by an NSArray, make the wrapping accessor methods use NSUInteger
for counts and indices as NSArray does.

+++ cocos2d/CCEffectRenderer.m
@@ -151,7 +151,7 @@
     [self freeAllRenderTargets];
     
     CCEffectRenderTarget *previousPassRT = nil;
-    for(int i = 0; i < effect.renderPassesRequired; i++)
+    for(NSUInteger i = 0; i < effect.renderPassesRequired; i++)
     {
         BOOL lastPass = (i == (effect.renderPassesRequired - 1));
         BOOL directRendering = lastPass && effect.supportsDirectRendering;
